requirements.txt
langchain-pinecone==0.2.2
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
flask==3.1.1
wtforms==3.2.1
flask-wtf==1.2.2
pdfplumber==0.11.6
cloudinary==1.44.0
email-validator==2.2.0
pymongo==4.13.0
python-dotenv==1.1.0
langchain==0.3.25
langchain-core==0.3.61
langchain-groq==0.3.2
langchain-huggingface==0.2.0